refactor: Include authentication in the module package id
