+++ json/src/androidTest/java/uk/gov/android/authentication/json/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.android.json
+package uk.gov.android.authentication.json
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
