+++ json/src/main/java/uk/gov/android/authentication/json/jwt/JwtVerifier.kt
@@ -1,4 +1,4 @@
-package uk.gov.android.json.jwt
+package uk.gov.android.authentication.json.jwt
 
 fun interface JwtVerifier {
     fun verify(encodedJsonWebToken: String, publicKeyJwkString: String): Boolean
